+++ final-project/src/components/home/Home.js
@@ -1,13 +1,31 @@
-import React from 'react';
-import './Home.css';
+import React, { useEffect, useState } from 'react';
 
-function Home() {
+const Home = () => {
+  const [posts, setPosts] = useState([]);
+
+  useEffect(() => {
+    const fetchPosts = async () => {
+      try {
+        const response = await fetch('http://localhost:8080/posts');
+        const posts = await response.json();
+        setPosts(posts);
+      } catch (err) {
+        console.error('Error fetching posts:', err);
+      }
+    };
+    fetchPosts();
+  }, []);
+
   return (
     <div>
-      <h1>Home</h1>
-      <p>Welcome to the home page!</p>
+      {posts.map((post) => (
+        <div key={post._id}>
+          <h3>{post.title}</h3>
+          <p>{post.content}</p>
+        </div>
+      ))}
     </div>
   );
-}
+};
 
 export default Home;
