+++ final-project/src/components/navbar/Navbar.js
@@ -5,14 +5,11 @@ import './Navbar.css';
 const Navbar = () => {
   return (
     <nav>
-      <ul>
-        <li><Link to="/">Home</Link></li>
-        <li><Link to="/trending">Trending</Link></li>
-        <li><Link to="/search">Search</Link></li>
-        <li><Link to="/account">Account</Link></li>
-        <li><Link to="/posts">Posts</Link></li>
-        <li><Link to="/create-post">Create Post</Link></li>
-      </ul>
+      <ul><Link to="/">Home</Link></ul>
+      <ul><Link to="/trending">Trending</Link></ul>
+      <ul><Link to="/search">Search</Link></ul>
+      <ul><Link to="/account">Account</Link></ul>
+      <ul><Link to="/create-post">Create Post</Link></ul>
     </nav>
   );
 };
