--- final-project/src/components/post/PostList.js
@@ -1,31 +1,0 @@
-import React, { useEffect, useState } from 'react';
-
-const PostList = () => {
-  const [posts, setPosts] = useState([]);
-
-  useEffect(() => {
-    const fetchPosts = async () => {
-      try {
-        const response = await fetch('http://localhost:8080/posts');
-        const posts = await response.json();
-        setPosts(posts);
-      } catch (err) {
-        console.error('Error fetching posts:', err);
-      }
-    };
-    fetchPosts();
-  }, []);
-
-  return (
-    <div>
-      {posts.map((post) => (
-        <div key={post._id}>
-          <h3>{post.title}</h3>
-          <p>{post.content}</p>
-        </div>
-      ))}
-    </div>
-  );
-};
-
-export default PostList;
